requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# OpenAI Integration
langchain-openai
openai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

pyttsx3

Flask
flask-cors

pygetwindow
paho-mqtt
pyautogui
pywhatkit
speedtest-cli
faiss-cpu
screen_brightness_control
pycaw
yt_dlp
opencv-python
dlib
face_recognition
resemblyzer
speechbrain